refactor surface modification algorithms
